Move numpyredis stuff into client.py
